Search elasticsearch Bro documents
Search elasticsearch Bro documents for bad (based on threat intel) URIs.  The input file is a list (line-feed delimited) of known malicious web sites.  This script checks against all documents with a field of 'uri' or 'query'.  The 'uri' and 'query' field is checked against ingested Bro documents from dns.log and http.log files.  An example of running this is:
 python searchBadURL.py -p 9200 -e 192.168.229.13 -f  urls.txt -i bro*
Where -p is the elasticsearch port. -e is the elasticsearch server. -f is the input file. -i is the index name.